Moving the pagination listener to the commons module
